chore(build): ensure that importing from src is not used inside /src directory
