+++ packages/dnb-eufemia/.eslintrc
@@ -39,13 +39,12 @@
     "no-restricted-imports": [
       "error",
       {
-        "paths": [
+        "patterns": [
           {
-            "name": "@dnb/eufemia",
+            "group": ["@dnb/eufemia/*"],
             "message": "Do not use recursive module @dnb/eufemia!"
           }
-        ],
-        "patterns": ["@dnb/eufemia/*"]
+        ]
       }
     ],
     "no-console": "off",
@@ -118,6 +117,26 @@
         "import/default": "off",
         "import/no-named-as-default-member": "off",
         "import/no-named-as-default": "off"
+      }
+    },
+    {
+      "files": ["**/src/**/*"],
+      "rules": {
+        "no-restricted-imports": [
+          "error",
+          {
+            "patterns": [
+              {
+                "group": ["@dnb/eufemia/*"],
+                "message": "Do not use recursive module @dnb/eufemia!"
+              },
+              {
+                "group": ["**/src/*"],
+                "message": "Do not import from src – but rather define correct relative paths!"
+              }
+            ]
+          }
+        ]
       }
     },
     {
